Add Phase 4 token extensions: i64 numbers, float literals, new keywords
- Change Token::Number(i32) to Token::Number(i64) and Expr::Number accordingly
- Add Token::Float(f64) for floating-point literals
- Add keywords: break, continue, as, nobreak
- Update Display impl for all new tokens
- Add lexer tests for new tokens

+++ src/parser/ast.rs
@@ -62,7 +62,7 @@ pub enum UnaryOp {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Expr {
-    Number(i32),
+    Number(i64),
     Ident(String),
     BinaryOp {
         op: BinOp,
